pyproject.toml
[project]
name = "confluencetonotion"
version = "0.7.0"
requires-python = ">= 3.10"
dependencies = [
    "confluence-markdown-exporter",
    "atlassian-python-api",
    "python-dotenv",
    "notion-markdown",
    "requests",
    # Phase 2: GraphRAG
    "neo4j",
    "neo4j-graphrag",
    "anthropic[bedrock]",
    "openai",
    "httpx",
    "pandas",
    "openpyxl",
    "fastapi",
    "uvicorn[standard]",
]

[project.scripts]
pipeline = "pipeline.run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["pipeline", "upload", "graphrag"]

[tool.uv.sources]
confluence-markdown-exporter = { path = "package/confluence-markdown-exporter", editable = true }
notion-markdown = { path = "package/notion-markdown", editable = true }
